// src/app/admin-view/models/mysociety.ts
export class MySociety {
    rId: number;
    sId: String;
    userMobileNo: string;
    refMobileNo: string;
    firstName: string;
    lastName: string;
    societyRegNo: string;
    roleId: string;
    imei: string;
    userSimSerialNo: string;
    address: string;
    state: string;
    district: string;
    taluka: string;
    pincode: string;
    otp: string;
    userPassword: string;
    token: string;
    sessionToken: string;
    status: string;
    newPassword: string;
    blockNo: string;
    flatNo: string;
    societyName: string;
    image: string;
    encryptData: string;
}